LayoutTests/http/tests/security/mixedContent/insecure-image-with-securecookie-block-expected.txt

CONSOLE MESSAGE: line 4: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image-secure-cookie-block.html was allowed to display insecure content from http://127.0.0.1:8080/security/resources/compass.jpg.

CONSOLE MESSAGE: line 6: secureCookie=yes
This test opens a window that tries to read a secure cookie after an insecure image has been loaded. This should allow reading of the secure cookie since only 'inactive' insecure content was loaded on this page.
